# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation for LIWACH system",
        "description": "API Documentation for LIWACH system",
        "contact": {
            "email": "edenabdisa90@yahoo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://liwachapi.herokuapp.com/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/membership": {
            "get": {
                "tags": [
                    "Membership"
                ],
                "summary": "Get list of membership",
                "description": "Returns list of membership",
                "operationId": "getMembershipList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Membership"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Membership"
                ],
                "summary": "Store new Membership",
                "description": "Returns Membership data",
                "operationId": "storeMembership",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MembershipRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Membership"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/membership/{id}": {
            "put": {
                "tags": [
                    "Membership"
                ],
                "summary": "Update existing project",
                "description": "Returns updated project data",
                "operationId": "updateMembership",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MembershipRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Membership"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Membership": {
                "title": "Membership",
                "description": "Membership model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "big integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the membership",
                        "type": "string",
                        "example": "normal,premium,gold"
                    },
                    "limit_of_post": {
                        "title": "Limit of post",
                        "description": "Limit of post that this memebrship is allowed",
                        "type": "integer",
                        "example": 100
                    },
                    "transaction_limit": {
                        "title": "Transation Limit",
                        "description": "Transation Limit in a day or in a week",
                        "type": "integer",
                        "example": 6
                    },
                    "timestamp": {
                        "title": "Timestamp",
                        "description": "This is the timestamp that we got when this data is added.",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                }
            },
            "MembershipRequest": {
                "title": "Membership",
                "description": "Membership model",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the membership",
                        "type": "string",
                        "example": "normal,premium,gold"
                    },
                    "limit_of_post": {
                        "title": "Limit of post",
                        "description": "Limit of post that this memebrship is allowed",
                        "type": "integer",
                        "example": 100
                    },
                    "transaction_limit": {
                        "title": "Transation Limit",
                        "description": "Transation Limit in a day or in a week",
                        "type": "integer",
                        "example": 6
                    }
                }
            },
...
        }
    }
}